requirements.txt
streamlit>=1.31.0
groq>=0.4.2
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4
pypdf>=3.17.0
python-docx>=1.1.0
numpy>=1.24.0
torch>=2.1.0
transformers>=4.36.0